update endpoints depending on dev environment
